removed templates
